core/bloom: move bit operation functions to utils
